backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.2.13
cloudinary==1.44.1
Pillow==11.3.0
